added some params to the modifier cards' functions, split ModifierCardsController into a logic and a rendering controller
